requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# NLP and embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Clustering and dimensionality reduction
umap-learn>=0.5.3
hdbscan>=0.8.33

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Data processing
tqdm>=4.65.0
pyyaml>=6.0

# Text processing
spacy>=3.5.0
nltk>=3.8

# Optional: for better PDF handling if needed later
# pypdf>=3.0.0
# pdfplumber>=0.9.0

# Development
jupyter>=1.0.0
pytest>=7.3.0
black>=23.0.0
